+++ packages/cli/src/__fixtures__/import-json/expect-sync/src/index.ts
@@ -1,0 +1,7 @@
+import { getTranslator } from "@hi18n/core";
+import { book } from "./locale";
+
+const { t } = getTranslator(book, "ja");
+// @ts-ignore
+t("example/greeting");
+t("example/greeting2");