+++ packages/cli/src/__fixtures__/import-json/expect-sync/src/locale/index.ts
@@ -1,0 +1,13 @@
+import { Message, Book } from "@hi18n/core";
+import catalogEn from "./en";
+import catalogJa from "./ja";
+
+export type Vocabulary = {
+  "example/greeting": Message;
+  "example/greeting2": Message;
+};
+
+export const book = new Book<Vocabulary>({
+  en: catalogEn,
+  ja: catalogJa,
+});